requirements.txt
aiogram==3.21.0
aiosqlite==0.21.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2
aiogram_i18n~=1.4
PyYAML~=6.0.2